separate search functionality into separate module
