+++ index.html
@@ -21,7 +21,7 @@
     <script type="text/javascript" src="https://cdn.anychart.com/releases/8.10.0/js/anychart-ui.min.js"></script>
     <script type="text/javascript" src="https://cdn.anychart.com/releases/8.10.0/js/anychart-ui-binding.min.js"></script>
     <script type="text/javascript" src="https://cdn.anychart.com/releases/8.10.0/js/anychart-data-adapter.min.js"></script>
-    <script type="text/javascript" src="./scripts/chart.js" defer></script>
+    <script type="module" src="./scripts/chart.js"></script>
 </head>
 
 <body>
